new FEAT BukuKolaborasiResource Filament

+++ app/Observers/BukuDijualObserver.php
@@ -13,11 +13,13 @@ class BukuDijualObserver
     public function created(buku_dijual $buku_dijual): void
     {
         $originalCover = $buku_dijual->getOriginal('cover_buku');
-        $originalCover = $buku_dijual->getOriginal('file_buku');
 
         if ($buku_dijual->isDirty('cover_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
         }
+
+        $originalCover = $buku_dijual->getOriginal('file_buku');
+
         if ($buku_dijual->isDirty('file_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
         }
@@ -29,11 +31,13 @@ class BukuDijualObserver
     public function updated(buku_dijual $buku_dijual): void
     {
         $originalCover = $buku_dijual->getOriginal('cover_buku');
-        $originalCover = $buku_dijual->getOriginal('file_buku');
 
         if ($buku_dijual->isDirty('cover_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
         }
+
+        $originalCover = $buku_dijual->getOriginal('file_buku');
+
         if ($buku_dijual->isDirty('file_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
         }
@@ -44,7 +48,7 @@ class BukuDijualObserver
      */
     public function deleted(buku_dijual $buku_dijual): void
     {
-        if (!is_null($buku_dijual->cover_buku) || !is_null($buku_dijual->file_buku) || !is_null($buku_dijual->storage_buku_dijual()->get('nama_generate'))) {
+        if (!is_null($buku_dijual->storage_buku_dijual()->get('nama_generate'))) {
             // get all the name_generate
             $name_generate = $buku_dijual->storage_buku_dijual()->get('nama_generate');
             // foreach name_generate
@@ -52,8 +56,13 @@ class BukuDijualObserver
                 // delete the file
                 Storage::disk('public')->delete($name['nama_generate']);
             }
+        }
 
+        if (!is_null($buku_dijual->cover_buku)) {
             Storage::disk('public')->delete($buku_dijual->cover_buku);
+        }
+
+        if (!is_null($buku_dijual->file_buku)) {
             Storage::disk('public')->delete($buku_dijual->file_buku);
         }
     }
@@ -71,6 +80,16 @@ class BukuDijualObserver
      */
     public function forceDeleted(buku_dijual $buku_dijual): void
     {
+        if (!is_null($buku_dijual->storage_buku_dijual()->get('nama_generate'))) {
+            // get all the name_generate
+            $name_generate = $buku_dijual->storage_buku_dijual()->get('nama_generate');
+            // foreach name_generate
+            foreach ($name_generate as $name) {
+                // delete the file
+                Storage::disk('public')->delete($name['nama_generate']);
+            }
+        }
+
         if (!is_null($buku_dijual->cover_buku)) {
             Storage::disk('public')->delete($buku_dijual->cover_buku);
         }
@@ -83,11 +102,12 @@ class BukuDijualObserver
     public function saved(buku_dijual $buku_dijual): void
     {
         $originalCover = $buku_dijual->getOriginal('cover_buku');
-        $originalCover = $buku_dijual->getOriginal('file_buku');
 
         if ($buku_dijual->isDirty('cover_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
         }
+
+        $originalCover = $buku_dijual->getOriginal('file_buku');
 
         if ($buku_dijual->isDirty('file_buku') && $originalCover !== null) {
             Storage::disk('public')->delete($originalCover);
